chore(fbw): move SignalBar to common component
