+++ plugins/lime-plugin-align/src/style.less
@@ -1,15 +1,3 @@
-.bar {
-	height: 0.4rem;
-}
-
-.signal {
-	display: flex;
-	flex-direction: column;
-	margin-left: auto;
-	font-size: 4rem;
-	justify-content: center;
-}
-
 .stationHostname {
 	font-size: 3rem;
 }
