+++ plugins/lime-plugin-fbw/src/style.less
@@ -2,25 +2,8 @@ button:disabled {
     opacity: 0.3;
 }
 
-.bar {
-	height: 0.4rem;
-}
-
-.signal {
-	display: flex;
-	flex-direction: column;
-	margin-left: auto;
-	font-size: 2.6rem;
-	justify-content: center;
-}
-
 .fetchingName {
 	font-size: 2rem;
-}
-
-.unit {
-	font-size: 0.3em;
-	margin-left: 0.1em;
 }
 
 .itemHidden{
