+++ src/components/signalbar/index.js
