+++ src/components/signalbar/style.less
@@ -1,0 +1,15 @@
+.bar {
+	height: 0.4rem;
+}
+
+.signal {
+	display: flex;
+	flex-direction: column;
+	margin-left: auto;
+	font-size: 4rem;
+	justify-content: center;
+}
+.unit {
+	font-size: 0.3em;
+	margin-left: 0.1em;
+}
